Use separate backend method for Vault create_proof
